backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
asyncpg
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
